addpic/requirements.txt
altgraph==0.15
cycler==0.10.0
future==0.16.0
macholib==1.9
matplotlib==2.1.1
numpy==1.14.0
pefile==2017.11.5
Pillow==5.0.0
pygame==1.9.3
PyInstaller==3.3.1
pyparsing==2.2.0
pypiwin32==220
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
wxPython==4.0.0b2
